- Moving api package to core, since it will be used on MovieActivity
